backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
pydantic==2.10.3
pydantic-settings==2.7.0
google-genai==1.7.0
numpy==2.2.1
alembic==1.14.0
httpx==0.28.1
beautifulsoup4==4.12.3
ddgs==7.5.3
